ekyc/requirements.txt
absl-py
aiohappyeyeballs
aiohttp
aiosignal
albucore
albumentations
annotated-types
anyio
argon2-cffi
argon2-cffi-bindings
attrs
blinker
certifi
cffi
charset-normalizer
click
coloredlogs
contourpy
cycler
Cython
easydict
ecdsa
python-jose
fastapi
filelock
Flask
flatbuffers
fonttools
frozenlist
fsspec
greenlet
grpcio
grpcio-tools
h11
httpcore
httpx
humanfriendly
idna
imageio
imutils
iniconfig
opt_einsum
packaging
pandas
pika
pillow
psycopg2-binary
pluggy
prettytable
propcache
protobuf
psutil
py-cpuinfo
pyasn1
pyclipper
pycparser
pycryptodome
pydantic
pydantic-settings
pydantic_core
Pygments
pyparsing
pytest
pytest-asyncio
pika
minio
fastapi
uvicorn
redis
sqlalchemy
sqlalchemy-utils
aiohttp
requests
python-magic
python-multipart
websockets
aio-pika